+++ lib/src/structures/client_options.dart
@@ -22,4 +22,15 @@ class ClientOptions {
     this.userAgent = "Fortnite/++Fortnite+Release-18.21-CL-17811397 Android/11",
     this.logLevel = Level.ALL,
   });
+
+  /// client options from json method
+  factory ClientOptions.fromJson(Map<String, dynamic> json) => ClientOptions(
+        log: json["log"] is bool ? json["log"] : false,
+        deviceAuth: json["deviceAuth"] is Map<String, dynamic>
+            ? DeviceAuth.fromJson(json["deviceAuth"])
+            : throw Exception(
+                "deviceAuth property is not a valid Map<String, dynamic>"),
+        logLevel: json["logLevel"] is Level ? json["logLevel"] : null,
+        userAgent: json["userAgent"] is String ? json["userAgent"] : null,
+      );
 }