+++ lib/src/structures/device_auth.dart
@@ -12,10 +12,20 @@ class DeviceAuth {
   /// display name for user of device auth
   late String displayName;
 
+  /// constructor
   DeviceAuth({
     required this.accountId,
     required this.deviceId,
     required this.secret,
     this.displayName = "",
   });
+
+  /// device auth from json method
+  factory DeviceAuth.fromJson(Map<String, dynamic> json) => DeviceAuth(
+        accountId: json["accountId"] is String ? json["accountId"] ?? "" : "",
+        deviceId: json["deviceId"] is String ? json["deviceId"] ?? "" : "",
+        secret: json["secret"] is String ? json["secret"] ?? "" : "",
+        displayName:
+            json["displayName"] is String ? json["displayName"] ?? "" : "",
+      );
 }