+++ CHANGELOG.md
@@ -89,3 +89,4 @@
 -   Ability to check campaign profile of other players.
 -   More MCP endpoints.
 -   Set supported creator method to common core profile.
+-   onSessionUpdate event stream on main client.